src/pyproject.toml
[tool.poetry]
name = "rentality"
version = "2.0.0"
description = ""
authors = ["rushil"]

[tool.poetry.dependencies]
python = ">=3.6,<3.7"
Django = "~=2.2.0"
celery = "~=4.4.0"
python-dotenv = "~=0.11.0"
django-autocomplete-light = "~=3.5.1"
django-debug-toolbar = "~=2.2.0"
django-cities = "~=0.6.0"
djangorestframework = "~=3.10.0"
djangorestframework-gis = "~=0.15.0"
django-allauth = "~=0.41.0"
easy-thumbnails = "~=2.7.0"
django-formtools = "~=2.2.0"
django-summernote = "~=0.8.11.0"
django-celery-results = "~=1.2.0"
django-celery-beat = "~=1.6.0"
psycopg2 = "~=2.8.4"
elasticsearch-dsl = "~=6.3"
elasticsearch = "~=6.3"
stripe = "~=2.42"
django-sendgrid-v5 = "~=0.8.1"
sentry-sdk = "~=0.14.1"
django-ipware = "^2.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
